requirements.txt
numpy<2.0.0
torch
tensorboard
wandb
numba
scipy
